refactor notes
