+++ src/App.jsx
@@ -6,6 +6,7 @@ import Main from './components/Main';
 import Friends from './components/Friends';
 import News from './components/News';
 import routes from './routes';
+import Bookmarks from './components/Bookmarks';
 
 const App = () => (
   <Switch>
@@ -13,6 +14,7 @@ const App = () => (
     <Route path={routes.login} component={Login} />
     <Route path={routes.friends} component={Friends} />
     <Route path={routes.news} component={News} />
+    <Route path={routes.bookmarks} component={Bookmarks} />
   </Switch>
 );
