+++ src/components/Main/Wall/index.js
@@ -21,7 +21,6 @@ import {
 import WallCreateArticle from '../WallCreateArticle';
 import FormStatus from './FormStatus';
 import BlockNotes from '../Articles/blockNotes/BlockNotes';
-import BlockComments from '../Articles/blockComment/BlockComments';
 
 const leftBlockItems = [
   'День рождения',
@@ -70,7 +69,6 @@ const Wall = () => (
     </WallInfoBlock>
     <WallCreateArticle />
     <BlockNotes />
-    <BlockComments />
   </WallContainer>
 );
